requirements.txt
#Python==3.9
pyyaml==6.0.2
schema==0.7.7
requests==2.32.3
beautifulsoup4==4.13.3
fake-headers==1.0.2

fastapi==0.115.12
uvicorn==0.34.2
sqlmodel==0.0.24

psycopg2==2.9.10
sqlalchemy~=2.0.40
bs4==0.0.2
starlette==0.46.2
pyjwt==2.10.1
passlib==1.7.4
pydantic~=2.11.3
exceptiongroup~=1.2.2
email_validator==2.2.0
bcrypt==4.0.1
streamlit==1.45.0

tensorflow==2.19.0
alembic==1.15.2
st-pages==1.0.1
numpy~=2.0.2
python-multipart==0.0.20